WIP: Put policy / CLI methods into rotator world
